♻️ Refactor :  useStyle에 있던 isResponsive, editResponsive 함수 새 파일로 분리
